Share resources between server and client
