+++ Client/Assets/entities/mobs/cow.js
@@ -1,7 +1,7 @@
 import { AI, Mob } from "../mob.js";
 import { Body, BodyPart } from "../../game/body.js";
-import { Items } from "../../game/items.js";
-import { LootItem, LootTable } from "../../game/lootTable.js";
+import { Items } from "@minecraftjs/shared/items.js";
+import { LootItem, LootTable } from "@minecraftjs/shared/lootTable.js";
 import { Sounds, playRandomSoundFromArray } from "../../game/sounds.js";
 import { Vector2 } from "../../utils/classes.js";
 import { BLOCK_SIZE } from "../../utils/globals.js";
