+++ Client/Assets/entities/mobs/pig.js
@@ -1,8 +1,8 @@
 import { runtime } from "../../utils/runtime.js";
 import { AI, Mob } from "../mob.js";
 import { Body, BodyPart } from "../../game/body.js";
-import { Items } from "../../game/items.js";
-import { LootItem, LootTable } from "../../game/lootTable.js";
+import { Items } from "@minecraftjs/shared/items.js";
+import { LootItem, LootTable } from "@minecraftjs/shared/lootTable.js";
 import { Sounds, playPositionalSound, playRandomSoundFromArray } from "../../game/sounds.js";
 import { Vector2 } from "../../utils/classes.js";
 import { BLOCK_SIZE } from "../../utils/globals.js";
