+++ Client/Assets/game/inventoryItem.js
@@ -1,62 +1,12 @@
 import { runtime } from "../utils/runtime.js";
-import { getItem } from "./item.js";
-import { inventoryRegistry } from "./inventoryRegistry.js";
+import { InventoryItem as SharedInventoryItem } from "@minecraftjs/shared/inventoryItem.js";
+import { getItem } from "@minecraftjs/shared/item.js";
+import { inventoryRegistry } from "@minecraftjs/shared/inventoryRegistry.js";
 import { drawImage, drawRect, drawText } from "../utils/renderer.js";
 import { getSpriteSize, getSpriteUrl } from "../utils/texturePackLoader.js";
 import { getBlock } from "../world/block.js";
 
-export class InventoryItem {
-    constructor({ blockId = null, itemId = null, count = 0, props = {} } = {}) {
-        this.blockId = blockId;
-        this.itemId = itemId;
-        this.count = count;
-        this.props = props;
-
-        this.init();
-    }
-
-    init() {
-        if (this.itemId !== null) {
-            const item = getItem(this.itemId);
-
-            if (item.durability) {
-                if (!this.hasProp("durability"))
-                    this.setProp("durability", item.durability);
-            }
-        }
-    }
-
-    addProps(props) {
-        this.props = { ...this.props, ...props };
-    }
-
-    removeProps(props) {
-        for (const prop in props) {
-            delete this.props[prop];
-        }
-    }
-
-    hasProp(prop) {
-        return this.props[prop] !== undefined;
-    }
-
-    getProp(prop) {
-        return this.props[prop];
-    }
-
-    setProp(prop, value) {
-        this.props[prop] = value;
-    }
-
-    removeProp(prop) {
-        delete this.props[prop];
-    }
-
-    isEmpty() {
-        if (this.count <= 0) return true;
-        return this.blockId === null && this.itemId === null;
-    }
-}
+export class InventoryItem extends SharedInventoryItem {}
 
 export class InventorySlot {
     constructor({
@@ -105,41 +55,31 @@ export class InventorySlot {
         const actualWidth = spriteSize.width || 16;
         const actualHeight = spriteSize.height || 16;
 
-        // Get the default cutoff for block
         let cutoff = 0;
         if (item.blockId) cutoff = getBlock(item.blockId).defaultCutoff || 0;
 
-        // Calculate the height to draw based on cutoff
         const drawHeight = actualHeight - cutoff * actualHeight;
-
-        // The amount to move the sprite down due to the cutoff
         const moveDown = cutoff * actualHeight;
 
         const baseDrawSize = 48;
         const scaleX = (baseDrawSize / actualWidth) * size;
         const scaleY = (baseDrawSize / actualHeight) * drawHeight * size;
 
-        // Frame animation logic based on time (globalFrame could be set elsewhere)
-        const frameCount = Math.floor(actualHeight / actualWidth); // Total number of frames in the sprite sheet
+        const frameCount = Math.floor(actualHeight / actualWidth);
 
         const animationSpeed = 2;
         const frame = Math.floor(runtime.globalFrame / animationSpeed) % frameCount;
 
-        // Calculate the crop Y offset for the current frame
-        const cropY = frame * actualWidth; // Adjust the Y position for the current frame
+        const cropY = frame * actualWidth;
+        const cropHeight = Math.min(actualWidth, actualHeight - cropY);
 
-        // Ensure we don't go out of bounds
-        const cropHeight = Math.min(actualWidth, actualHeight - cropY); // Max height for each frame
-
-        // Set crop to match the current animation frame
         const crop = {
             x: 0,
-            y: moveDown + cropY, // Apply the moveDown offset for cutoff
+            y: moveDown + cropY,
             width: actualWidth,
             height: cropHeight,
         };
 
-        // Draw the sprite with animation applied through the crop
         drawImage({
             url: spritePath,
             x: slotX,
@@ -147,11 +87,10 @@ export class InventorySlot {
             scale: Math.min(scaleX, scaleY),
             centerX: false,
             dark: item.props.wall === true,
-            fixAnimation: false, // No longer need to fix to 16x16
+            fixAnimation: false,
             crop: crop,
         });
 
-        // Draw durability bar
         if (item.hasProp?.("durability")) {
             const durability = item.getProp("durability");
             const itemDef = getItem(item.itemId);
