+++ Server/package.json
@@ -3,6 +3,7 @@
     "version": "1.0.0",
     "type": "module",
     "dependencies": {
+        "@minecraftjs/shared": "file:../Shared",
         "ws": "^8.18.1"
     }
 }
