+++ Shared/dimensions.js
@@ -1,0 +1,6 @@
+/** Static dimension ids (mutable active state stays on the client). */
+
+export const Dimensions = Object.freeze({
+    Overworld: 0,
+    Nether: 1,
+});
