+++ Shared/index.js
@@ -1,0 +1,24 @@
+/** Shared headless game data — usable from Client and Server. */
+
+export * from "./constants.js";
+export * from "./math.js";
+export { default as tooloud } from "./tooloud.js";
+export * from "./noise.js";
+export * from "./dimensions.js";
+export * from "./sounds.js";
+export * from "./blockType.js";
+export * from "./blockRegistry.js";
+export * from "./treesRegistry.js";
+export * from "./inventoryRegistry.js";
+export * from "./item.js";
+export * from "./items.js";
+export * from "./lootTable.js";
+export * from "./blocks.js";
+export * from "./trees.js";
+export * from "./biome.js";
+export * from "./structure.js";
+export * from "./structures.js";
+export * from "./chestLoot.js";
+export * from "./inventoryItem.js";
+export * from "./recipe.js";
+export * from "./recipes.js";
