+++ Shared/inventoryRegistry.js
