+++ Shared/math.js
@@ -1,4 +1,4 @@
-/** Pure math helpers with no game imports (avoids registry load cycles). */
+/** Pure math helpers with no game imports. */
 
 export function randomRange(min, max) {
     return Math.floor(Math.random() * (max - min)) + min;
